메인 켄버스 연결
