+++ .idea/.idea.seemarket-client/.idea/contentModel.xml
@@ -10,7 +10,6 @@
         <e p="GoogleARCore" t="Include">
           <e p="Editor" t="Include">
             <e p="ARCoreAndroidKeylessDependencies.template" t="Include" />
-            <e p="ARCoreAndroidKeylessDependencies.xml" t="Include" />
             <e p="ARCoreiOSDependencies.template" t="Include" />
           </e>
           <e p="Examples" t="Include">
@@ -357,6 +356,11 @@
             <e p="DrinkDetailCanvasControl.cs" t="Include" />
           </e>
           <e p="DrinkObject.cs" t="Include" />
+          <e p="MainCanvas" t="Include">
+            <e p="CameraScript.cs" t="Include" />
+            <e p="CanvasControl.cs" t="Include" />
+            <e p="UserInterface.cs" t="Include" />
+          </e>
           <e p="Model" t="Include">
             <e p="Drink.cs" t="Include" />
             <e p="Response.cs" t="Include" />
@@ -371,16 +375,18 @@
             <e p="SlotService.cs" t="Include" />
             <e p="StallService.cs" t="Include" />
           </e>
+          <e p="ShelfObject.cs" t="Include" />
+          <e p="Shelving.cs" t="Include" />
           <e p="ShelvingTest.cs" t="Include" />
           <e p="Singleton" t="Include">
+            <e p="CCameraManager.cs" t="Include" />
             <e p="CObjectPool.cs" t="Include" />
             <e p="CSingleton.cs" t="Include" />
+            <e p="CWebData.cs" t="Include" />
           </e>
           <e p="SlotObject.cs" t="Include" />
           <e p="TestCode.cs" t="Include" />
         </e>
-        <e p="ShelfObject.cs" t="Include" />
-        <e p="Shelving.cs" t="Include" />
         <e p="SocketIO" t="Include">
           <e p="JSONObject" t="Include">
             <e p="Editor" t="Include">
@@ -484,10 +490,12 @@
             <e p="WebSocketStream.cs" t="Include" />
           </e>
         </e>
-        <e p="test" t="Include">
-          <e p="CameraScript.cs" t="Include" />
-          <e p="CanvasControl.cs" t="Include" />
-          <e p="UserInterface.cs" t="Include" />
+        <e p="VSCode" t="Include">
+          <e p="Plugins" t="Include">
+            <e p="Editor" t="Include">
+              <e p="VSCode.cs" t="Include" />
+            </e>
+          </e>
         </e>
       </e>
       <e p="obj" t="ExcludeRecursive" />
