+++ Assets/Resources/MainCanvas.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 6133e9682b38a19488e439f77c9f0ab5
-folderAsset: yes
-DefaultImporter:
+guid: 3340b3823281c4d008c4722a2b9896f2
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
